+++ LocalPackages/Waitlist/Sources/Waitlist/Views/ActivityIndicator.swift
@@ -1,0 +1,40 @@
+//
+//  ActivityIndicator.swift
+//  DuckDuckGo
+//
+//  Copyright © 2022 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import SwiftUI
+import UIKit
+
+public struct ActivityIndicator: UIViewRepresentable {
+
+    public typealias UIViewType = UIActivityIndicatorView
+
+    public let style: UIActivityIndicatorView.Style
+
+    public init(style: UIActivityIndicatorView.Style) {
+        self.style = style
+    }
+
+    public func makeUIView(context: UIViewRepresentableContext<ActivityIndicator>) -> UIActivityIndicatorView {
+        return UIActivityIndicatorView(style: style)
+    }
+
+    public func updateUIView(_ view: UIActivityIndicatorView, context: UIViewRepresentableContext<ActivityIndicator>) {
+        view.startAnimating()
+    }
+}
