+++ LocalPackages/Waitlist/Sources/Waitlist/Views/InviteCodeView.swift
@@ -1,0 +1,52 @@
+//
+//  InviteCodeView.swift
+//  DuckDuckGo
+//
+//  Copyright © 2022 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import SwiftUI
+
+public struct InviteCodeView: View {
+
+    public let title: String
+    public let inviteCode: String
+
+    public init(title: String, inviteCode: String) {
+        self.title = title
+        self.inviteCode = inviteCode
+    }
+
+    public var body: some View {
+        VStack(spacing: 4) {
+            Text(title)
+                .font(.proximaNova(size: 17))
+                .foregroundColor(.white)
+                .padding([.top, .bottom], 4)
+
+            Text(inviteCode)
+                .font(.system(size: 34, weight: .semibold, design: .monospaced))
+                .padding([.leading, .trailing], 18)
+                .padding([.top, .bottom], 6)
+                .foregroundColor(.black)
+                .background(Color.white)
+                .cornerRadius(4)
+        }
+        .padding(4)
+        .background(Color.waitlistGreen)
+        .cornerRadius(8)
+    }
+
+}
